Implemented game mechanics.
Implemented game mechanics. Game is now fully working correctly. Will now focus on modifying resizing behaviour for JFXSheet and GameWindow. JFXSheet currently broken as it is in the process of being changed to update Label size when resizing. GameWindow is not resizable at the moment since the focus was on implementing the game mechanics and testing them only.