Improved game mechanics and added more themes
Improved look and mechanics of the atom trivia. When users choose a wrong answer the right answer will flash green for a two seconds before going into the next question. The app now takes into consideration streaks. If  users enter three (for level 3) correct answers in a row they get an extra life. Added 10 more atoms that can appear in level 1 as it was too repetitive.

Also added the option to choose themes for the look of the application from the settings. Users can now select three different looks for the application.